GPU: Reorganize some files, split OCL code in kernel and non-kernel related parts
